requirements/base.txt
appnope==0.1.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
crawler==0.1.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.1.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nest-asyncio==1.5.4
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-version==0.0.2
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
